# Improve the performance of StockBulkUpdate mutation