--- a/saleor/plugins/tests/sample_plugins.py
+++ b/saleor/plugins/tests/sample_plugins.py
@@ -255,6 +255,9 @@ class PluginSample(BasePlugin):
     def promotion_ended(self, promotion: "Promotion", previous_value: Any):
         return None
 
+    def product_variant_stock_updated(self, stock, previous_value: Any):
+        return None
+
     def get_checkout_line_tax_rate(
         self,
         checkout_info: "CheckoutInfo",